Reverted addition of submodules
The Jekyll code in the submodule 'gurjeet' started confusing the Github
Pages infrastructure and complain with errors, and cause HTTP 404 error
message on singh.im domain's homepage.

The submodules were added to represent the hierarchy in the DNS. But
this was not necessary, and since now it's causing problems, it's wise
to just get rig of this superfluous arrangement. The domain singh.im now
publishes its plain index files, as expected. And the subdomains `sggs`
and `gurjeet` continue to funciton as before because of their own
repositories being published via Github Pages.